+++ replacer-backend/replacer-repository/src/main/java/es/bvalero/replacer/page/count/PageCountCacheRepository.java
@@ -1,6 +1,6 @@
 package es.bvalero.replacer.page.count;
 
-import static java.time.temporal.ChronoUnit.MINUTES;
+import static java.time.temporal.ChronoUnit.HOURS;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.LoadingCache;
@@ -38,7 +38,7 @@ class PageCountCacheRepository implements PageCountRepository {
     // as modifications in the cache may happen while the count query is in progress.
     // We could lock the cache while querying but the query takes too long.
     // In theory this cache is mirroring the database reality so in the future the duration could be even longer.
-    private final Duration refreshTime = Duration.of(30, MINUTES);
+    private final Duration refreshTime = Duration.of(1, HOURS);
     private final LoadingCache<WikipediaLanguage, Map<StandardType, Integer>> counts = Caffeine
         .newBuilder()
         .refreshAfterWrite(refreshTime)