+++ replacer-backend/replacer-repository/src/main/java/es/bvalero/replacer/replacement/ReplacementSaveCacheRepository.java
@@ -4,7 +4,6 @@ import es.bvalero.replacer.common.domain.StandardType;
 import es.bvalero.replacer.common.domain.WikipediaLanguage;
 import es.bvalero.replacer.page.count.PageCountRepository;
 import java.util.Collection;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -27,19 +26,6 @@ class ReplacementSaveCacheRepository implements ReplacementSaveRepository {
 
     @Override
     public void add(Collection<IndexedReplacement> replacements) {
-        // In case of batch indexing the replacements may belong to several pages
-        // Also the same may contain several replacements of the same type
-        replacements
-            .stream()
-            .collect(
-                Collectors.toMap(
-                    IndexedReplacement::getPageKey,
-                    IndexedReplacement::getType,
-                    (existing, replacement) -> existing
-                )
-            )
-            .forEach((pageKey, type) -> pageCountRepository.increment(pageKey.getLang(), type));
-
         replacementSaveRepository.add(replacements);
     }
 
@@ -51,19 +37,6 @@ class ReplacementSaveCacheRepository implements ReplacementSaveRepository {
 
     @Override
     public void remove(Collection<IndexedReplacement> replacements) {
-        // In case of batch indexing the replacements may belong to several pages
-        // Also the same may contain several replacements of the same type
-        replacements
-            .stream()
-            .collect(
-                Collectors.toMap(
-                    IndexedReplacement::getPageKey,
-                    IndexedReplacement::getType,
-                    (existing, replacement) -> existing
-                )
-            )
-            .forEach((pageKey, type) -> pageCountRepository.decrement(pageKey.getLang(), type));
-
         replacementSaveRepository.remove(replacements);
     }