+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Make Admin menu in frontend a dropdown
 - Fix: reload empty cache to find pages left to review after
 - Consider replacements with the same context but far enough as different
+- Fix: update page count cache when indexing by type and not by replacement
 
 ## 2.23.6 (8 May 2023)
 - Fix regression: dropdown to change language